test(network): add some tests about Packet handling

+++ tests/server/packet/serializer/SerializerTests.cpp
@@ -20,7 +20,7 @@ protected:
 
 TEST_F(BigEndianSerializationTest, SerializeIntProducesCorrectBytes) {
     BigEndianSerialization serializer;
-    int value = 0x12345678;
+    uint64_t value = 0x12345678;
     std::vector<uint8_t> expected = {0x12, 0x34, 0x56, 0x78};
 
     std::vector<uint8_t> result = serializer.serializeInt(value);
@@ -30,7 +30,7 @@ TEST_F(BigEndianSerializationTest, SerializeIntProducesCorrectBytes) {
 
 TEST_F(BigEndianSerializationTest, SerializeLongProducesCorrectBytes) {
     BigEndianSerialization serializer;
-    int64_t value = 0x0123456789ABCDEF;
+    uint64_t value = 0x0123456789ABCDEF;
     std::vector<uint8_t> expected = {0x01, 0x23, 0x45, 0x67, 0x89, 0xAB, 0xCD, 0xEF};
 
     std::vector<uint8_t> result = serializer.serializeLong(value);
@@ -40,7 +40,7 @@ TEST_F(BigEndianSerializationTest, SerializeLongProducesCorrectBytes) {
 
 TEST_F(BigEndianSerializationTest, SerializeShortProducesCorrectBytes) {
     BigEndianSerialization serializer;
-    uint16_t value = 0xABCD;
+    uint64_t value = 0xABCD;
     std::vector<uint8_t> expected = {0xAB, 0xCD};
 
     std::vector<uint8_t> result = serializer.serializeShort(value);
@@ -50,7 +50,7 @@ TEST_F(BigEndianSerializationTest, SerializeShortProducesCorrectBytes) {
 
 TEST_F(BigEndianSerializationTest, SerializeCharProducesCorrectBytes) {
     BigEndianSerialization serializer;
-    uint8_t value = 0x7F;
+    uint64_t value = 0x7F;
     std::vector<uint8_t> expected = {0x7F};
 
     std::vector<uint8_t> result = serializer.serializeChar(value);
@@ -61,9 +61,9 @@ TEST_F(BigEndianSerializationTest, SerializeCharProducesCorrectBytes) {
 TEST_F(BigEndianSerializationTest, DeserializeIntProducesCorrectValue) {
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0x12, 0x34, 0x56, 0x78};
-    uint32_t expected = 0x12345678;
+    uint64_t expected = 0x12345678;
 
-    uint32_t result = serializer.deserializeInt(data);
+    uint64_t result = serializer.deserializeInt(data);
 
     EXPECT_EQ(result, expected);
 }
@@ -72,7 +72,7 @@ TEST_F(BigEndianSerializationTest, DeserializeIntWithInsufficientDataReturnsZero
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0x12, 0x34};
 
-    uint8_t result = serializer.deserializeInt(data);
+    uint64_t result = serializer.deserializeInt(data);
 
     EXPECT_EQ(result, 0);
 }
@@ -80,9 +80,9 @@ TEST_F(BigEndianSerializationTest, DeserializeIntWithInsufficientDataReturnsZero
 TEST_F(BigEndianSerializationTest, DeserializeLongProducesCorrectValue) {
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0x01, 0x23, 0x45, 0x67, 0x89, 0xAB, 0xCD, 0xEF};
-    int64_t expected = 0x0123456789ABCDEF;
+    uint64_t expected = 0x0123456789ABCDEF;
 
-    int64_t result = serializer.deserializeLong(data);
+    uint64_t result = serializer.deserializeLong(data);
 
     EXPECT_EQ(result, expected);
 }
@@ -91,7 +91,7 @@ TEST_F(BigEndianSerializationTest, DeserializeLongWithInsufficientDataReturnsZer
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0x01, 0x23, 0x45};
 
-    int64_t result = serializer.deserializeLong(data);
+    uint64_t result = serializer.deserializeLong(data);
 
     EXPECT_EQ(result, 0);
 }
@@ -99,9 +99,9 @@ TEST_F(BigEndianSerializationTest, DeserializeLongWithInsufficientDataReturnsZer
 TEST_F(BigEndianSerializationTest, DeserializeShortProducesCorrectValue) {
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0xAB, 0xCD};
-    uint16_t expected = 0xABCD;
+    uint64_t expected = 0xABCD;
 
-    uint16_t result = serializer.deserializeShort(data);
+    uint64_t result = serializer.deserializeShort(data);
 
     EXPECT_EQ(result, expected);
 }
@@ -109,9 +109,9 @@ TEST_F(BigEndianSerializationTest, DeserializeShortProducesCorrectValue) {
 TEST_F(BigEndianSerializationTest, DeserializeCharProducesCorrectValue) {
     BigEndianSerialization serializer;
     std::vector<uint8_t> data = {0x7F};
-    uint8_t expected = 0x7F;
+    uint64_t expected = 0x7F;
 
-    uint8_t result = serializer.deserializeChar(data);
+    uint64_t result = serializer.deserializeChar(data);
 
     EXPECT_EQ(result, expected);
 }
